+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `PasswordSettings::clear_words()`.
+- `PasswordSettings::remove_word_at()`.
 - `range_inc_from_str()` for getting a `RangeInclusive` from a `String`.
 
 ### Changed